use relative path for an image, rename card.jsx -> Card.jsx & update imports
